requirements.txt
Jinja2==2.7.3
Twisted==14.0.2
Werkzeug==0.9.6
argparse==1.2.1
cffi==0.8.6
characteristic==14.1.0
cryptography==0.5.4
klein==0.2.3
lxml==3.4.0
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
service-identity==14.0.0
six==1.8.0
git+https://github.com/cwaldbieser/treq.git@custom_agent
wsgiref==0.1.2
zope.interface==4.1.1
python-dateutil==2.2
